фикс парсинга погоды

+++ ARgronom/ARgronom/Models/Weather/WeatherApiResponse.cs
@@ -10,7 +10,7 @@ namespace ARgronom.Models.Weather
         public double lat { get; set; }
         public double lon { get; set; }
         public string timezone { get; set; }
-        public int timezone_offset { get; set; }
+        public double timezone_offset { get; set; }
         public Current current { get; set; }
         public List<Minutely> minutely { get; set; }
         public List<Hourly> hourly { get; set; }
@@ -19,7 +19,7 @@ namespace ARgronom.Models.Weather
 
     public class Weather
     {
-        public int id { get; set; }
+        public double id { get; set; }
         public string main { get; set; }
         public string description { get; set; }
         public string icon { get; set; }
@@ -27,27 +27,27 @@ namespace ARgronom.Models.Weather
 
     public class Current
     {
-        public int dt { get; set; }
-        public int sunrise { get; set; }
-        public int sunset { get; set; }
+        public double dt { get; set; }
+        public double sunrise { get; set; }
+        public double sunset { get; set; }
         public double temp { get; set; }
         public double feels_like { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double dew_point { get; set; }
         public double uvi { get; set; }
-        public int clouds { get; set; }
-        public int visibility { get; set; }
+        public double clouds { get; set; }
+        public double visibility { get; set; }
         public double wind_speed { get; set; }
-        public int wind_deg { get; set; }
+        public double wind_deg { get; set; }
         public double wind_gust { get; set; }
         public List<Weather> weather { get; set; }
     }
 
     public class Minutely
     {
-        public int dt { get; set; }
-        public int precipitation { get; set; }
+        public double dt { get; set; }
+        public double precipitation { get; set; }
     }
 
     public class Rain
@@ -57,17 +57,17 @@ namespace ARgronom.Models.Weather
 
     public class Hourly
     {
-        public int dt { get; set; }
+        public double dt { get; set; }
         public double temp { get; set; }
         public double feels_like { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double dew_point { get; set; }
         public double uvi { get; set; }
-        public int clouds { get; set; }
-        public int visibility { get; set; }
+        public double clouds { get; set; }
+        public double visibility { get; set; }
         public double wind_speed { get; set; }
-        public int wind_deg { get; set; }
+        public double wind_deg { get; set; }
         public double wind_gust { get; set; }
         public List<Weather> weather { get; set; }
         public double pop { get; set; }
@@ -94,22 +94,22 @@ namespace ARgronom.Models.Weather
 
     public class Daily
     {
-        public int dt { get; set; }
-        public int sunrise { get; set; }
-        public int sunset { get; set; }
-        public int moonrise { get; set; }
-        public int moonset { get; set; }
+        public double dt { get; set; }
+        public double sunrise { get; set; }
+        public double sunset { get; set; }
+        public double moonrise { get; set; }
+        public double moonset { get; set; }
         public double moon_phase { get; set; }
         public Temp temp { get; set; }
         public FeelsLike feels_like { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double dew_point { get; set; }
         public double wind_speed { get; set; }
-        public int wind_deg { get; set; }
+        public double wind_deg { get; set; }
         public double wind_gust { get; set; }
         public List<Weather> weather { get; set; }
-        public int clouds { get; set; }
+        public double clouds { get; set; }
         public double pop { get; set; }
         public double uvi { get; set; }
         public double? rain { get; set; }
